Fix partial application example
The examples misse a 'ref' in the function declaration resulting
in it defaulting to 'box'. This causes an error when writing to
the objects field.

The apply method in Bar creates a Foo that defaults to 'iso' and
gives an error about aliasing. Providing a type declaration defaults
it to 'ref'.