Add initial Gitbook support
Switching the tutorial off of Github page hosted tutorial and over
to Gitbook appears to have a few advantages.

1. Better navigation
2. Navigation isn't directly tied to organization of content
3. PDFs!
4. Mobi and EPub
5. More user control over font size and theme

If we wanted to switch back, its relatively minor so we are
moving forward with the switch. There's a number of improvements
to be made once we are on Gitbook but this feels like the right thing
to do.

Content will remain in Markdown (although we could switch to asciidoc
if we wanted to/needed more powerful formatting). We'll lose some
control over syntax highlighting in the short term, but we can get it
back if we get Pony formatting into highlight.js proper.