Add documentation for `as`
This is initial `as` documentation. It focuses on the two uses of
`as`: to cast objects as more specific types, and to specify a type
for items in a literal array. Two code examples are given for casting,
one involving union types and the other inheritance; there is one
example for specifying the type for items in an array. These
particular examples were used because they are short and somewhat
representative of the types of problems that can be solved with `as`.

This fixes issue #64.