Work around GitBook rendering error
Code blocks for unrecognized languages (and only for them)
don't correctly render '<=' characters. It seems to be
treating as html and removing everything after them so

while count <= 10 do

becomes

while count

This commit changes

```pony

to

```

to address in the one known case.

There are other uses of <= in the tutorial but they are in
blocks that don't trigger.