Baby step towards preserving pre gitbook urls
When serving via mkdocs/github pages, all content
that was in "docs" was served as /. When we moved
to gitbook, that started being served as /docs/
thereby breaking the old url. I've started
to fix that by moving the content up to the top level
of the repo.

I'm not doing it in one swoop because doing a section
at a time makes it easier for me to review each section
and make sure that the organization, naming etc makes
sense and abides by new content organization rules.

Leaving things I haven't done that for in docs makes it
easy to keep track of progress.