Clean up initial installation explanation
Fixes <code> bugs in the intro to installing the Pony compiler -- previously text was interspersed with the code snippets, so I switched them to be in-line code instead for better readability. This also adds a link directly to the "Building ponyc from source" section since it is a long way down the page in the current page organization.