Add notes about overloaded functions
Several places in the tutorial stated that Pony does not support
overloaded functions. While possibly true for some definition of the
term "overloaded function", this caused some confusion because Pony
does support case functions, which achieve several of the goals of
overloaded functions.

This change adds links to the case function documentation in places
case functions might be helpful, and adds a section to the case
function documentation discussing the ways that case functions might
differ from ones expectations of overloaded functions.

Fixes #106