Point installation directions at the GH repo README
Previously we were maintaining instructions inline in
the tutorial. This was great if you had downloaded a
PDF or even if you were using the tutorial on the web
as you didn't have to jump to a different site. However,
it also meant maintaining two copies of the installation
instructions. The tutorial directions would regularly
get out of date. From a maintenance standpoint, it was
a nightmare.