Difference in assignation between variable and field
My tests have shown that assignation works a bit differently between variables and fields. This is what I'm trying to explained in this update of the field section of the tutorial. Also, I've put a bunch of example because I find the tutorial usually miss examples. And sometimes, nothing beats a good example for understanding :)