Clarify results of integer and floating point division by zero
The division by zero section is now more clear about the fact that
integer division by 0 is 0 and floating point division by zero is
`inf` or `-inf`. This well help developers understand what to expect
from division and make it clear that the results will differ depending
on whether the division is of integers or floats.

Closes #310