Remove the first-person wording of viewpoint adaptation.
While it is true that access to receivers and fields will follow
this, in that `this` is just a normal variable, I feel this outlook
could be confusing when understood with receiver recovery, and
changing of capabilities.

A reader might conclude or speculate that a variable with type iso,
can only see its own contents with those viewpoints, but the notion
of an object accessing itself is most closely aligned with calling
a function (or behavior). Since this will almost certainly be
`fun ref` and `fun box`, the table will not apply here, which
could obfuscate learning.