Describe capabilities of fields before discussing viewpoint adaptation (#447)
This is a simultaneous issue I had when starting and a proposed fix with the knowledge I have now.

When first reading this, the section on viewpoints and fields was rather confusing, since it didn't feel clear what it meant to have a capability of a field (and we never actually mentioned them in the classes chapter). This introduces them, and makes sure to specify that a field makes the same exact global guarantees.