The mandatory FFI "use" change improves our safety story. (#475)
Even though it is true that we can't make guarantees around FFI calls, the pony type system coupled with the now mandatory "use" statements does provide all the tools a programmer needs to guarantee that the data that they pass into their FFI functions is correct.

(Where "correct" is left as an exercise to the reader as they are the ones that need to strictly define what is "correct".  We can't automatically validate either the "shape" of the data or derive who manages the memory).

The point being, if they correctly describe the data using pony types, the pony compiler *will* enforce the shape of that data for them consistently throughout their application.  That is *huge*.

(How to map pony-types and build type-systems that are flexible _and_ safe is 95% of the difficulty of doing C-FFI)