Elaborate on trait vs interface differences
This commit provides a rewrite of the traits and interfaces documentation.
The goals are:

- Be more nuanced in the differences between trait and interface

In particular, acknowledge that interfaces can be used for both
structural and nominal subtyping.

- Explain the differences between trait and interface

The previous content was focused almost entirely on nominal vs
structural subtyping and didn't give folks who aren't experienced
with both examples to demonstrate the advantages of each approach
as well as the different affordances they provide.